requirements.txt
Flask>=2.3.0,<3.0.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-Login>=0.6.0,<1.0.0
Flask-WTF>=1.1.0,<2.0.0
WTForms>=3.0.0,<4.0.0
Werkzeug>=2.3.0,<3.0.0
sentence-transformers>=2.2.0,<3.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.2.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
PyPDF2>=3.0.0,<4.0.0
python-docx>=0.8.11,<1.0.0
gunicorn>=20.0.0,<22.0.0 